requirements.txt
Django<=2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
django-cors-headers==3.1.1
factory_boy==2.12.0